pyproject.toml
[project]
name = "halo-bridge"
version = "0.1.0"
description = "Sync articles from Halo 2.x blog to CSDN, cnblogs, and modb.pro"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27,<1.0",
    "click>=8.1,<9.0",
    "pyyaml>=6.0,<7.0",
    "Markdown>=3.5,<4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
]

[project.scripts]
halo-bridge = "halo_bridge.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
